Settings tests completed, added space to toString in Settings
