--- OrganizationApp/src/tests/Settings.java
@@ -1,34 +1,0 @@
-package tests;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class Settings {
-
-	@Test
-	public void testGetUserName() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testSetUserName() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testGetUserEmail() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testSetUserEmail() {
-		fail("Not yet implemented");
-	}
-
-	@Test
-	public void testToString() {
-		fail("Not yet implemented");
-	}
-
-}
